fix: Models can't be initialized in parallel
Running multiple application instances in parallel (eg. for testing) will cause the the await table() to come back in non-deterministic time, when `initialized` may or may not have already been set. Checking for model.store is a simple check to fail out of the rest of the init if the model has already been initialized by the time the await resolves.